:truck: refactor: diectory refactor
