--- src/main/java/com/jokim/sivillage/api/best/product/application/BestProductService.java
@@ -1,8 +1,0 @@
-package com.jokim.sivillage.api.best.product.application;
-
-import com.jokim.sivillage.api.best.product.dto.BestProductResponseDto;
-import java.util.List;
-
-public interface BestProductService {
-    List<BestProductResponseDto> getBestProduct();
-}
