+++ src/main/java/com/jokim/sivillage/api/trending/mostViewProduct/application/MostViewProductService.java
@@ -1,0 +1,8 @@
+package com.jokim.sivillage.api.trending.mostViewProduct.application;
+
+import com.jokim.sivillage.api.trending.mostViewProduct.dto.MostViewProductDto;
+import java.util.List;
+
+public interface MostViewProductService {
+    List<MostViewProductDto> getMostViewProduct();
+}
