+++ src/main/java/com/jokim/sivillage/api/trending/product/presentation/BestProductController.java
@@ -1,20 +1,15 @@
-package com.jokim.sivillage.api.best.product.presentation;
+package com.jokim.sivillage.api.trending.product.presentation;
 
 
-import com.jokim.sivillage.api.best.product.application.BestProductService;
-import com.jokim.sivillage.api.best.product.dto.BestProductResponseDto;
-import com.jokim.sivillage.api.best.product.vo.BestProductResponseVo;
-import com.jokim.sivillage.api.customer.dto.in.SignUpRequestDto;
-import com.jokim.sivillage.api.customer.vo.in.SignUpRequestVo;
+import com.jokim.sivillage.api.trending.product.application.BestProductService;
+import com.jokim.sivillage.api.trending.product.dto.BestProductResponseDto;
+import com.jokim.sivillage.api.trending.product.vo.BestProductResponseVo;
 import com.jokim.sivillage.common.entity.BaseResponse;
-import com.jokim.sivillage.common.entity.BaseResponseStatus;
 import io.swagger.v3.oas.annotations.Operation;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
